ai-engine/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Authentication & Security
pyjwt==2.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# HTTP clients
httpx==0.25.2
aiohttp==3.9.1

# Database drivers
asyncpg==0.29.0
motor==3.3.2
redis==5.0.1

# AI/ML Libraries
openai==1.3.7
anthropic==0.7.8
google-generativeai==0.3.2
cohere==4.37

# Vector stores and embeddings
pinecone-client==2.2.4
chromadb==0.4.18
weaviate-client==3.25.3
sentence-transformers==2.2.2

# Data processing
numpy==1.24.4
pandas==2.1.4

# Monitoring and metrics
prometheus-client==0.19.0

# Utilities
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1